fix(run-data-sync.yml): correct typo in step name from 'parqent' to 'parquet' fix(run-data-sync.yml): remove unnecessary pdm install command as it is already installed in the workflow fix(run-data-sync.yml): simplify the run command for the 'Push new runs' step
The typo in the step name 'parqent' was corrected to 'parquet' to ensure accuracy. The unnecessary `pdm install duckdb` command was removed as it is already installed in the workflow, streamlining the process. Additionally, the run command for the 'Push new runs' step was simplified for better readability and maintainability.